agents/orchestrator/pyproject.toml
[project]
name = "orchestrator-agent"
version = "0.1.0"
description = "orchestrator Agent "
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.108.0",
    "google-adk>=1.10.0",
    "pydantic-settings>=2.8.1",
    "tabulate>=0.9.0",
    "google-auth>=2.36.0",
    "requests>=2.32.3",
    "llama-index>=0.12",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator-agent"]

[tool.uv]
dev-dependencies = []

